requirements.txt
absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
databricks-sdk==0.30.0
Deprecated==1.2.14
docker==7.1.0
einops==0.8.0
entrypoints==0.4
filelock==3.13.1
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.65.4
gunicorn==22.0.0
h5py==3.12.1
huggingface-hub==0.24.5
humanfriendly==10.0
humanize==4.11.0
idna==3.4
imageio==2.34.1
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lpips==0.1.4
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
MedPy==0.5.2
mlflow==2.15.1
mlflow-skinny==2.15.1
monai==1.3.2
mpmath==1.3.0
networkx==3.2.1
nibabel==5.2.1
numpy==1.26.3
omegaconf==2.3.0
onnxruntime==1.18.1
opencv-python==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.4
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.28.1
rich==13.9.2
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.32
sqlparse==0.5.1
sympy==1.12
tempo @ file:///home/Code/Tempo
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.8
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
TorchIntegral @ file:///home/Code/SSRNO/_TO
torchio==0.20.1
torchvision==0.16.1+cu121
tqdm==4.66.4
triton==2.1.0
typer==0.12.5
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.20.0
